house-cleaning: changed interface naming scheme
When we are finally going to be used in production the on hover captions
may confuse the user into thinking that we have a universal Props
interface that represents all components. This may lead to usage
confusion and may be counter-intuitive for a purposes. In this commit, I
also changed:
- Modal tests to include WCAG 2.0 compliance
- Removed unecessary console logging from test files