foundational: syncing cleaned history
After having cleaned the entire commit history, I went file by file and
copy pasted the contents from the latest commit of the old history and
formatted the code. This allows us to catch any lost or gained code from
the clean up and makes sure that the latest version works the exact same
way as it used to.