component: added area variant to Button
Added a new variant called area in Button component. This variant does not have style of a button, but works exactly as a button.

deleted the changes of button group to
sync with main and fixed area variant

component: added area variant to Button

Added a new variant called area in Button component. This variant does not have style of a button, but works exactly as a button.

merged with main

component: added area variant to Button

Added a new variant called area in Button component. This variant does not have style of a button, but works exactly as a button.

merged with main