foundational: changed apollo identification process
Previously, we were using `name` and `displayName` which didn't work in
production due the minification process. In order to counter-act this
problem all Apollo components have gained a new non-changeable property:
`data-apollo`. This property contains the name of the component and
since its a string literal with direct impact to the DOM element, this
property shouldn't be modified during the production process.

Refs:
https://mparavano.medium.com/find-filter-react-children-by-type-d9799fb78292