foundational: refactored format children and types
Improved quality of life in Apollo by determining clear types for
handling Overloaded components. This resulted in a type refactor for
FormatChildren and all components who use it. This update introduces the
following new types:

- Interface: Used for interfacing components with optional parent
  properties
- RenderAll: A function type that describes the method that implements
  formatted children. Used for interfacing components that by default
  implement passthrough like the View and Section
- ParentProps: a definitive type for defining what parent props really
  are. This is now implemented by the Interface and Overload type.

Additionally, a slight change to FormatChildren includes a standalone
parameter for the `children` property and now the `componentMap` goes
before the `parentProp` paremeter.