component: revamped table component
This commit aimed on tackling several challeneges. For one, we are now
able to create a meta-component structure thanks to the creation of the
`ChildrenMeta` class. This enables us to gather property data from
components, in this case the `Table`, and use it for render purposes
all in one render. This has allowed us to revamp and improve the `Table`
both functionally, and dynamically. Furthermore, the table is now both
visually and unit tested for use in Apollo