util:  added theming to Apollo (#89)
This commit was primarily focused around including theming of the most
high priority components in Apollo. This lead to the creation of the
class object `ApolloCSS` and the interface `ApolloTheme` both these
combined allow us to add a styling utility that has the option of being
either server-side or client-side preferred while remaining static. This
commit also fixes the `Select` component behavior to act more like
expected by our users.