component: updated dropdown and image components (#90) (#91)
The main changes that were made to Apollo include the addition of
indicative arrows to all variants of `Dropdown` components. This is to
be more inline with a previous issue with the UI. Additionally, we
updated the `Image` component to use the new `Spinner` component instead
of having its own custom element embedded within.

Other changes in this PR include:

- Updated Theming to not overwrite any other component style changes
- Updated theming documentation
- Added `label` prop to `Spinner` in order to overwrite default
  aria-labeling
- Removed unnecessary styling in `Image`