util: added deep format and helper (#92)
This commit was dedicated to creating an alternate solution to
`FormatChildren` which would look throug the **entire** DOM tree to find
Apollo components to overload. This solution is the first step to
tackling abstraction issues we were having with forms that would prevent
live validation of deeply nested children. Alongside the new Apollo
component `HandleOverloads`, `DeepFormat` can even look into components
that are user-created to retain functionality.

Other changes:
* removed unnecessary destructure of `Button` properties