foundational: added user-wrapped input validation to `Form`
This commit does 2 big things that were missing in Apollo. A way to read
Apollo components that were wrapped by user-defined components and then
implementing that solution into the `Form` so we have features like live
validation and action-data population. Previously we defined that
`DeepFormat` would handle this use-case but what ended up happening is
that by introducing a feature that was computationaly more intense, we
ended up creating a potential pitfall for development.

This is why we went back and merged the functionalities of `DeepFormat`
and `FormatChildren`. Meaning, we are going to stick to using Apollo
components to create layout and perpatuate processing, however, now we
add the ability to read into user-defined components to check if
they have an input that requires live validation.

Finally, we improved the error registering for `TextInput` components to
be more sensitive to re-render based on the type of error brought.