feature: added calendar component
In this commit, I focused primarily on adding a calendar component to
Apollo that meets the minimum requirements that parallel the current
calendar in Health/FedCiv platforms. It has the major functionalities
and can capture when there is a date change, as well as marking
important dates, etc. I also added two new props to `Section`, being
`inline` and `gap`, which allow the flexbox to decide whether it wants
to be `inline-flex` or have a `gap` between children.