feature create calendar dropdown (#103)
* parent 70bde612747f7a6c3c3787c1ee1ed8b2fb45c2f4
author Eddie Lopez <ejlopez00@gmail.com> 1689864114 -0400
committer Eddie Lopez <ejlopez00@gmail.com> 1689887586 -0400

feature: created date select component

In this commit I focused on creating a text input with the ability to
dropdown a Calendar. This input allows the user to select a date and see
it displayed in a text input. Additionally, they may type a date and see
its result reflected in the input.