This. Is. Sparta. (#111)
* component: updated dropdown and image components (#90)

The main changes that were made to Apollo include the addition of
indicative arrows to all variants of `Dropdown` components. This is to
be more inline with a previous issue with the UI. Additionally, we
updated the `Image` component to use the new `Spinner` component instead
of having its own custom element embedded within.

Other changes in this PR include:

- Updated Theming to not overwrite any other component style changes
- Updated theming documentation
- Added `label` prop to `Spinner` in order to overwrite default
  aria-labeling
- Removed unnecessary styling in `Image`

* util: added deep format and helper (#92)

This commit was dedicated to creating an alternate solution to
`FormatChildren` which would look throug the **entire** DOM tree to find
Apollo components to overload. This solution is the first step to
tackling abstraction issues we were having with forms that would prevent
live validation of deeply nested children. Alongside the new Apollo
component `HandleOverloads`, `DeepFormat` can even look into components
that are user-created to retain functionality.

Other changes:
* removed unnecessary destructure of `Button` properties

* feature: updated form to find inputs in user-defined components (#94)

This commit does 2 big things that were missing in Apollo. A way to read
Apollo components that were wrapped by user-defined components and then
implementing that solution into the `Form` so we have features like live
validation and action-data population. Previously we defined that
`DeepFormat` would handle this use-case but what ended up happening is
that by introducing a feature that was computationaly more intense, we
ended up creating a potential pitfall for development.

This is why we went back and merged the functionalities of `DeepFormat`
and `FormatChildren`. Meaning, we are going to stick to using Apollo
components to create layout and perpatuate processing, however, now we
add the ability to read into user-defined components to check if
they have an input that requires live validation.

Finally, we improved the error registering for `TextInput` components to
be more sensitive to re-render based on the type of error brought.

* feature: added main branch deployments (#95)

This commit is geared towards making automatic deployments everytime there is a merge to the main branch

* bugfix: added tags filter to deployment

* release: v0.1.38

This release introduces the following to Apollo:

- Users will now be able to delegate Apollo processing to their own
  components using `OverloadHandler` and `Overloader`
- Improved Unit testing
- New documentation
- Automatic deployments in main branch (yet-to-be tested)

* feature: added calendar component (#97)

In this commit, I focused primarily on adding a calendar component to
Apollo that meets the minimum requirements that parallel the current
calendar in Health/FedCiv platforms. It has the major functionalities
and can capture when there is a date change, as well as marking
important dates, etc. I also added two new props to `Section`, being
`inline` and `gap`, which allow the flexbox to decide whether it wants
to be `inline-flex` or have a `gap` between children.

* feature: added date range capabilities to calendar

In this update, I just reinforced what we previously had with the `Calendar` component by adding a date range variant.

* release: v0.1.42

Update now introduces the long await calendar with date range
capabilities. Additionally, we introduce some nice feature upgrades and
such. Nothing too crazy.

* feature: added value to calendar (#100)

* foundational: dropping testing support for node v12.22

* release: v0.1.43

This release just introduces some minor tweaks and a much needed feature
to the existing Calendar component. It also goes back and fixes some
testing errors we were having.

* feature: create the accordion component (#102)

feature: added accordion component

In this commit, I added an important component for one of our latest
projects, Capture. This component, of course, is the Accordion. This is
a button with an expandable container that has the ability to declutter
functional and informational data.

* feature create calendar dropdown (#103)

* parent 70bde612747f7a6c3c3787c1ee1ed8b2fb45c2f4
author Eddie Lopez <ejlopez00@gmail.com> 1689864114 -0400
committer Eddie Lopez <ejlopez00@gmail.com> 1689887586 -0400

feature: created date select component

In this commit I focused on creating a text input with the ability to
dropdown a Calendar. This input allows the user to select a date and see
it displayed in a text input. Additionally, they may type a date and see
its result reflected in the input.

* Apollo is now Sparta (#105)

* Changed Apollo to Sparta

* changed apollo to sparta in components and css.

* Apollo is now Sparta

* v1.0 (#108)

* v1.0

* Dropping testing support for node v14

* Consolidate `main` and `dev` branches (#110)

* Apollo V1.0.0 (#104)

* updated branch version

* Added loading prop to Image component for lazy (#106)

* lockfile

* Fixed merge conflicts, bumped version to 1.2.0, finished switching names to sparta

* Lock react-hook-form version

---------

Co-authored-by: Eduardo Lopez <52265974+elopez00@users.noreply.github.com>
Co-authored-by: Darius-Deese <138823553+Darius-Deese@users.noreply.github.com>

* Fix merge conflicts with main (#112)

* Apollo V1.0.0 (#104)

* updated branch version

* Added loading prop to Image component for lazy (#106)

---------

Co-authored-by: Eduardo Lopez <52265974+elopez00@users.noreply.github.com>
Co-authored-by: Darius-Deese <138823553+Darius-Deese@users.noreply.github.com>

---------

Co-authored-by: Eddie Lopez <ejlopez00@gmail.com>
Co-authored-by: Eduardo Lopez <52265974+elopez00@users.noreply.github.com>
Co-authored-by: Md Ahmed Reza Siddique <ahmedreza80@gmail.com>
Co-authored-by: Darius-Deese <138823553+Darius-Deese@users.noreply.github.com>